src/pyproject.toml
[project]
name = "etl"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "pandas>=2.2.3",
    "pyarrow>=19.0.1",
    "fastparquet>=2024.11.0",
    "boto3>=1.37.28",
    "python-dotenv",
    "requests",
    "web3>=7.13.0"
]